Hello-Back!
Hello!
I just saw your post on FreeCodeCamp - but it's 1:00 am here so I was a little late for a reply.
I'm learning too so I thought I would try to use github to send you a reply. I've never made a "pull request" before and I'm not all that familiar with git operations in general, but hey, there's a first time for everything, right?
Anyway, I have to get back to work now. I just wanted to learn a little bit more about github and send you a quick note to say "hang in there!"
~Jason